Moved some of the functions out of app.js into separate utils component files
