+++ src/utils/getPlaylist.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+function getPlaylist() {
+  let playlist = JSON.parse(localStorage.getItem("playlist"));
+  if (playlist) {
+    return playlist;
+  } else {
+    return [];
+  }
+}
+
+export default getPlaylist;
