+++ src/utils/savePlaylist.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function savePlaylist(playlist) {
+  localStorage.setItem("playlist", JSON.stringify(playlist));
+}
+
+export default savePlaylist;
